Made cannon.sh and cannon.asm more portable.
cannon.sh
--
Added STATUS flag.
Checks for the assembler chosen.
Checks that md5sum is there.
Checks that an emulator is reachable.
If either of these three checks fails, the script exits.
Otherwise it assembles, runs md5sum check, and starts up the emulator.

Cannon/cannon.asm
--
Changed the path to the includes to reflect the changes in cannon.sh